+++ changelog.md
@@ -1,1 +1,5 @@
-- fixed an issue with fake level using more resources than it needs, generating empty chunks when it could have had just 1 
+- **CompatWoodType**
+  - Added **Environmental**'s leaves & Fixed [EveryCompat's #769](https://github.com/MehVahdJukaar/WoodGood/issues/769)
+  - Added **Aether Genesis**'s leaves  - @VoidLeech
+  - Added one more leaves from **Ancient Aether** - @VoidLeech
+  - Updated **Unusual End**'s chorus_cane